Renamed the app 'timetable_requirements' to be 'data_upload'
