+++ timetable/interfaces/data_upload/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class TimetableRequirementsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'interfaces.timetable_requirements'
+    name = 'interfaces.data_upload'
